refactor(domain): migra campos quantidade de BigDecimal para Integer
Campos quantidade em OsItem, MovimentacaoEstoque e SaldoEstoque migraram de BigDecimal(NUMERIC 10,3) para Integer, eliminando zeros desnecessarios na serializacao JSON (ex: 1.000 -> 1). Campos monetarios precoUnitario/valorTotal/subtotal permanecem BigDecimal. DTOs de request/response, repository, services, DataLoader e todos os testes (110 no total) atualizados e validados.

+++ src/main/java/br/com/fiap/oficina/domain/repository/MovimentacaoEstoqueRepository.java
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -24,5 +23,5 @@ public interface MovimentacaoEstoqueRepository extends JpaRepository<Movimentaca
         FROM MovimentacaoEstoque m
         WHERE m.produto.id = :produtoId
     """)
-    BigDecimal calcularSaldoPorProduto(@Param("produtoId") Long produtoId);
+    Integer calcularSaldoPorProduto(@Param("produtoId") Long produtoId);
 }
